+++ lib/td_audit/application.ex
@@ -3,36 +3,16 @@ defmodule TdAudit.Application do
 
   use Application
 
-  alias TdAudit.Repo
-  alias TdAuditWeb.Endpoint
-
   # See https://hexdocs.pm/elixir/Application.html
   # for more information on OTP Applications
   def start(_type, _args) do
-    import Supervisor.Spec
+    env = Application.get_env(:td_audit, :env)
 
-    notifications_worker = %{
-      id: TdAudit.NotificationLoader,
-      start: {TdAudit.NotificationLoader, :start_link, []}
-    }
-
-    # Define workers and child supervisors to be supervised
-    children = [
-      # Start the Ecto repository
-      supervisor(Repo, []),
-      # Start the endpoint when the application starts
-      supervisor(Endpoint, []),
-      # Start your own worker by calling: TdAudit.Worker.start_link(arg1, arg2, arg3)
-      # worker(TdAudit.Worker, [arg1, arg2, arg3]),
-      supervisor(Exq, []),
-      %{
-        id: TdAudit.CustomSupervisor,
-        start:
-          {TdAudit.CustomSupervisor, :start_link,
-           [%{children: [notifications_worker], strategy: :one_for_one}]},
-        type: :supervisor
-      }
-    ]
+    children =
+      [
+        TdAudit.Repo,
+        TdAuditWeb.Endpoint
+      ] ++ children(env)
 
     # See https://hexdocs.pm/elixir/Supervisor.html
     # for other strategies and supported options
@@ -43,7 +23,23 @@ defmodule TdAudit.Application do
   # Tell Phoenix to update the endpoint configuration
   # whenever the application is updated.
   def config_change(changed, _new, removed) do
-    Endpoint.config_change(changed, removed)
+    TdAuditWeb.Endpoint.config_change(changed, removed)
     :ok
+  end
+
+  defp children(:test), do: [
+    {TdAudit.Broadway, broadway_config()}
+  ]
+
+  defp children(_env) do
+    [
+      TdAudit.Notifications.Loader,
+      TdAudit.Subscriptions.EventHandler,
+      {TdAudit.Broadway, broadway_config()}
+    ]
+  end
+
+  defp broadway_config do
+    Application.get_env(:td_audit, TdAudit.Broadway)
   end
 end