+++ priv/repo/migrations/20200602085037_alter_events_timestamps.exs
@@ -1,0 +1,15 @@
+defmodule TdAudit.Repo.Migrations.AlterEventsTimestamps do
+  use Ecto.Migration
+
+  def change do
+    alter table("events") do
+      modify(:ts, :utc_datetime_usec, from: :utc_datetime)
+      modify(:inserted_at, :utc_datetime_usec, from: :utc_datetime)
+      remove(:updated_at, :utc_datetime)
+    end
+
+    alter table("subscriptions") do
+      modify(:last_consumed_event, :utc_datetime_usec, from: :utc_datetime)
+    end
+  end
+end